Implemented a basic vi_view struct with clipping
I'm not quite happy yet with how the rendering hierarchy is
structured. I originally wanted vi_view to contain the buffer it was
rendering for, but, because the buffer needs to render to a vi_view
itself, the includes would be recursive. In addition, the CPU usage
for this is pretty high (around 30%), so I might have to investigate that.