Added font styling
I'm not so sure if I'm happy with how this works right now. I've had
to remove the const qualifier on the pointer returned by
vi_font_store_ensure_font in order to allow the user to set the font,
but I'm not really sure if that's the best way of doing it. If someone
doesn't clear the style before using some font, they might get styling unexpectedly.