Added a split widget
Due to how the widgets are structured, we're not able to access the
bounds of the widget. This makes the events a bit weird to
handle. As for the key presses and scrolling, we can make vi_state
store a selected widget. This might remove all of the widget event
functions in the vtable. As for the click events, I'll probably need
to add either a bounds function to the vtable, or add the bounds field to
the struct itself.